+++ plugin-hrm-form/src/components/tabbedForms/TabbedForms.tsx
@@ -13,7 +13,7 @@ import Case from '../case';
 import { namespace, contactFormsBase, routingBase, RootState, configurationBase } from '../../states';
 import { updateCallType, updateForm } from '../../states/contacts/actions';
 import { searchResultToContactForm } from '../../services/ContactService';
-import { completeContactlessTask } from '../../services/formSumbissionHelpers';
+import { completeContactlessTask } from '../../services/formSubmissionHelpers';
 import { changeRoute } from '../../states/routing/actions';
 import type { TaskEntry } from '../../states/contacts/reducer';
 import type { TabbedFormSubroutes } from '../../states/routing/types';
